pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "AI Agent with MCP support"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.23",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.3.1",
    "langchain-groq>=0.3.2",
    "truststore>=0.10.1",
    "mcp>=1.0.0",
    "requests>=2.32.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"